requirements.txt
appdirs==1.4.4
autopep8==1.5.5
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
flask-paginate==0.8.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.0
kbbi==0.4.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
nltk==3.5
numpy==1.20.1
pandas==1.2.2
protobuf==3.15.7
pycodestyle==2.6.0
PySastrawi==1.2.0
pyTelegramBotAPI==3.7.6
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
soupsieve==2.1
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.56.0
urllib3==1.26.3
visitor==0.1.3
Werkzeug==1.0.1
